+++ components/src/App.tsx
@@ -1,7 +1,9 @@
 import React from "react";
 import { Modal } from "../../components/src/lib/Modal/Modal";
 import "./App.css";
-import AlertModal from "./lib/AlarmModal/AlarmModal";
+import { AlertModal } from "./lib/AlarmModal/AlarmModal";
+import { ConfirmModal } from "./lib/ConfirmModal/ConfirmModal";
+import { PromptModal } from "./lib/PromptModal/PromptModal";
 
 function App() {
   return (
@@ -22,9 +24,27 @@ function App() {
       <Modal.Provider>
         <Modal.Trigger>AlertModal 열기</Modal.Trigger>
 
-        <AlertModal title="Alert" size="medium" modalPosition="center" closeButtonPosition="top">
+        <AlertModal title="Alert" size="small" modalPosition="center" closeButtonPosition="top">
           AlertModal 내용
         </AlertModal>
+      </Modal.Provider>
+
+      {/* PromptModal */}
+      <Modal.Provider>
+        <Modal.Trigger>PromptModal 열기</Modal.Trigger>
+
+        <PromptModal title="Prompt" size="medium" modalPosition="center" closeButtonPosition="top">
+          PromptModal 내용
+        </PromptModal>
+      </Modal.Provider>
+
+      {/* ConfirmModal */}
+      <Modal.Provider>
+        <Modal.Trigger>ConfirmModal 열기</Modal.Trigger>
+
+        <ConfirmModal title="Confirm" size="small" modalPosition="center" closeButtonPosition="top">
+          ConfirmModal 내용
+        </ConfirmModal>
       </Modal.Provider>
     </>
   );
