+++ components/src/lib/AlarmModal/AlarmModal.tsx
@@ -31,5 +31,3 @@ export const AlertModal: React.FC<
     </>
   );
 };
-
-export default AlertModal;
