+++ components/src/lib/index.ts
@@ -1,1 +1,4 @@
-export { Modal } from "../lib/Modal/Modal";
+export { AlertModal } from "./AlarmModal/AlarmModal";
+export { ConfirmModal } from "./ConfirmModal/ConfirmModal";
+export { Modal } from "./Modal/Modal";
+export { PromptModal } from "./PromptModal/PromptModal";
